Add files via upload
The Muscle Studio is my first self-made project of a front-end gym website using HTML,CSS. In this webpage, I have inserted various navigation buttons about the information of the gym customer is looking for, nearby gym with details of its equipment. Moreover, a dropdown box which contains the details of the client (such as name, address, age, contact info., etc.). Currently, this is a very basic webpage but in upcoming days I would be uploading more webpages with different designs and new features.